Move `applyBattlerTags` to `utils` subfolder
